clean up and start of antelope
Started antelope,  in midst of setting up behaviour tasks and services.

+++ Ava/Source/Ava/RoutePoint.h
@@ -14,7 +14,7 @@ class AVA_API ARoutePoint : public ATargetPoint
 	GENERATED_BODY()
 public:
 		UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = RouteID)
-		float RouteNumber;
+		int RouteNumber;
 	
 	
 };
